Implement TECH-736 - check for xbash ransomware
The xbash ransomware drops all databases and leaves behind
a "PLEASE_READ_ME" database with instructions on how to pay
the ransom. Check if this directory exists via get_mysql_datadir()
and warn. This also ensured the upadte of all_malware_checks()
so the --csi arg would trigger this check too.