TECH-676: Add check_root_dns_resolvers()
Ideally, we would not use bash in Perl, but there isn't a great way to
ask a specific DNS server without pulling in a new module. That would be
fine but we are also in the middle of moving to a new DNS module. Added
TODO line for future update to module once 84 more widely accepted.