Increment version from v4.99.209 to v4.99.210
Forgot to increase the version in the last commit. Ensure that never
happens again by updating MakeFile with a version check. The version
check only acts on branches beginning with 'v', i.e. the version
branches. If the VERSION variable in ssp doesn't match the branch
version, the make fails.